backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
django-csp==3.8
nemoguardrails==0.8.0
